reorganize analysis functions, make a waveform basis example
