+++ smalldata_tools/ana_funcs/waveformFunc.py
@@ -2,7 +2,7 @@ import tables
 import numpy as np
 import scipy
 from scipy import signal as scipy_signal
-from DetObject import DetObjectFunc
+from smalldata_tools.DetObject import DetObjectFunc
 from smalldata_tools.roi_rebin import spectrumFunc
 from smalldata_tools.utilities import templateArray as utility_templateArray
